+++ Src/PINT_Common/PINT_CommonPhysX3_Mesh.cpp
@@ -113,7 +113,7 @@ udword MeshObjectManager::AddObject(void* object)
 	else
 	{
 		Index = mFirstFree;
-		mFirstFree = udword(mObjects[Index]);
+		mFirstFree = udword((size_t)mObjects[Index]);
 		mObjects[Index] = object;
 	}
 
@@ -138,7 +138,7 @@ bool MeshObjectManager::DeleteObject(void* object, const udword* index)
 	if(object && mObjects[Index]!=object)
 		return false;
 
-	mObjects[Index] = reinterpret_cast<void*>(mFirstFree);
+	mObjects[Index] = reinterpret_cast<void*>((size_t)mFirstFree);
 	mFirstFree = Index;
 	return true;
 }
